Use debian as base image and bash for scripts
Alpine is much smaller but it suffers from severe DNS issues. It lacks
support for search domains and tries to resolve a dns name via all
configure nameservers and returns the first answer. Since we keep a
non-mesos dns server for backup, sometimes this one is faster and we get
NXDOMAIN for internal names.

- https://github.com/gliderlabs/docker-alpine/issues/8
- https://github.com/gliderlabs/docker-alpine/issues/11