Optionally make sysctl setting a soft failure.
By default, if `systcl` fails it results in a hard failure. In other
words, the main `run` script will exit and the container will fail.

By setting `HAPROXY_SYSCTL_NONSTRICT` to a non-empty string, the
`sysctl` failures will be treated as soft failures.

This was introduced to workaround in issue where Docker will mount /proc
within the container as read-only, even when the container is running in
privileged mode.