Refactor #607 (#630)
* Fallback on next Marathon instance on exception

If requests raise a TimeoutException it might be due to the current
instance being down. We should fallback to the next instance

* Stick to the host with the last successful request

If we make a requests to a marathon host and it failed it might be
because the host is down. We shouldn't keep trying making request to
this host and instead requesting the last host that answer correctly.

* marathon-lb: refactor host fallback